Simplified and Greenified
tried simplifying the code by cutting out things that didn't seem necessary, for instance, the image that was advised against having and the complex navigation that went along with it.  Also changed the general color vide from purple to green.  Finally, I've started work on making it better match the resume that's been settled upon.